Revert "Attempt to fix broken treeIndex without botchering performance on iOS"
This reverts commit e3117f6f82a3e8fa0c7276891c18cafc6f55bf98.

+++ viewer/packages/rendering/src/glsl/base/updateFragmentColor.glsl
@@ -23,7 +23,7 @@ vec3 packNormalToRgb( const in vec3 normal ) {
 }
 
 void updateFragmentColor(
-    int renderMode, vec4 color, int treeIndex, 
+    int renderMode, vec4 color, float treeIndex, 
     vec3 normal, float depth, sampler2D matCapTexture, 
     int geometryType) {
     if (renderMode == RenderTypeColor || renderMode == RenderTypeEffects) {
